+++ src/main/java/com/hibit2/hibit2/auth/controller/AuthController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.hibit2.hibit2.auth.dto.TokenRequest;
-import com.hibit2.hibit2.auth.dto.AccessAndRefreshTokenResponse;
+import com.hibit2.hibit2.auth.dto.request.TokenRequest;
+import com.hibit2.hibit2.auth.dto.response.AccessAndRefreshTokenResponse;
 import com.hibit2.hibit2.auth.service.AuthService;
 import com.hibit2.hibit2.global.dto.CommonResponse;
