+++ src/main/java/com/hibit2/hibit2/auth/dto/response/AccessTokenResponse.java
@@ -1,0 +1,16 @@
+package com.hibit2.hibit2.auth.dto.response;
+
+public class AccessTokenResponse {
+    private String accessToken;
+
+    private AccessTokenResponse() {
+    }
+
+    public AccessTokenResponse(final String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public String getAccessToken() {
+        return accessToken;
+    }
+}
