+++ src/test/java/com/hibit2/hibit2/member/domain/MemberRepositoryTest.java
@@ -6,9 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import com.hibit2.hibit2.member.domain.Member;
-import com.hibit2.hibit2.member.domain.MemberRepository;
-import com.hibit2.hibit2.member.domain.SocialType;
+import com.hibit2.hibit2.member.repository.MemberRepository;
 
 @DataJpaTest
 class MemberRepositoryTest {
